+++ Poco.Evolved.SQL/Database/DatabaseType.cs
@@ -10,7 +10,7 @@ namespace Poco.Evolved.SQL.Database
     public enum DatabaseType
     {
         /// <summary>
-        /// A generic dialect using standard SQL. This dialect may not work with every database.
+        /// A generic dialect using ANSI SQL. This dialect may not work with every database.
         /// </summary>
         Generic,
 
@@ -20,22 +20,22 @@ namespace Poco.Evolved.SQL.Database
         Firebird,
 
         /// <summary>
-        /// A dialect for Microsoft SQL Server
+        /// A dialect for Microsoft SQL Server. (only generic support for now)
         /// </summary>
         MSSQLServer,
 
         /// <summary>
-        /// A dialect for MySQL.
+        /// A dialect for MySQL. (only generic support for now)
         /// </summary>
         MySQL,
 
         /// <summary>
-        /// A dialect for Oracle.
+        /// A dialect for Oracle. (only generic support for now)
         /// </summary>
         Oracle,
 
         /// <summary>
-        /// A dialect for PostgreSQL.
+        /// A dialect for PostgreSQL. (only generic support for now)
         /// </summary>
         PostgreSQL,