+++ ReleaseNotes.md
@@ -12,7 +12,7 @@
 #### Bugfixes
 * Use a transaction for creating the table (or equivalent for non-SQL databases) for installed versions
 
-###v_._._ (upcoming release)
+### vX.X.X (upcoming release)
 #### Features
 * Targets .NET Standard 1.3, .NET Standard 2.0 and .NET 4.5
 
@@ -39,6 +39,6 @@
   * SQLite
 * Targets .NET Standard 2.0 and .NET 4.5
 
-###v_._._ (upcoming release)
+### vX.X.X (upcoming release)
 #### Features
 * Targets .NET Standard 1.3, .NET Standard 2.0 and .NET 4.5